backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
typing-extensions>=4.7.0
python-dotenv>=1.0.0
httpx>=0.24.0
pandas>=2.0.0
numpy>=1.24.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.1
pytest>=7.3.1
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.2.0
